backend/requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
websockets==12.0

# Database
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
ldap3==2.9.1

# File Handling
python-multipart==0.0.9

# Validation
email-validator==2.1.0.post1

# Development Tools
pytest==8.0.0
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Logging
python-json-logger==2.0.7

# WebSockets
python-dotenv==1.0.1
pydantic-settings==2.1.0
websockets==12.0